Add MoveNetPose class
