+++ Assets/Samples/MoveNet/MoveNetPose.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3d6be79f2802d41d891e76c854af7aa8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
